pyproject.toml
[tool.poetry]
name = "visualize-surveys"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.3"
django-ordered-model = "^3.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
